requirements.txt
django-heroku
gunicorn
git+https://github.com/yuce/pyswip@master#egg=pyswip
